Document upload integration. Fixed Meetings Creation issue.

+++ backend/backend-app/backend-app/DTOs/DocumentDTO.cs
@@ -6,7 +6,7 @@
         public string FileName { get; set; } 
         public string Description { get; set; } 
         public string MimeType { get; set; } 
-        public float FileSizeBytes { get; set; }
+        public int FileSizeBytes { get; set; }
         public DateTime UploadedAt { get; set; }
     }
 }
